linearAlgebra/animations/pyproject.toml
[tool.poetry]
name = "animations"
version = "0.1.0"
description = ""
authors = ["Sayan Naskar <nascarsayan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
manim = "^0.19.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
column_and_null_spaces = "column_and_null_spaces:main"